fix: make filter method generic
